fix(server): Fix method to get full uri with proxy headers
