+++ pkg/s3-proxy/server/utils/utils.go
@@ -6,6 +6,7 @@ import (
 	"html/template"
 	"net/http"
 	"path/filepath"
+	"strings"
 
 	"github.com/Masterminds/sprig"
 	"github.com/oxyno-zeta/s3-proxy/pkg/s3-proxy/config"
@@ -173,5 +174,51 @@ func GetRequestURI(r *http.Request) string {
 		scheme = "https"
 	}
 
-	return fmt.Sprintf("%s://%s%s", scheme, r.Host, r.URL.RequestURI())
+	return fmt.Sprintf("%s://%s%s", scheme, RequestHost(r), r.URL.RequestURI())
+}
+
+func RequestHost(r *http.Request) string {
+	// not standard, but most popular
+	host := r.Header.Get("X-Forwarded-Host")
+	if host != "" {
+		return host
+	}
+
+	// RFC 7239
+	host = r.Header.Get("Forwarded")
+	_, _, host = parseForwarded(host)
+
+	if host != "" {
+		return host
+	}
+
+	// if all else fails fall back to request host
+	host = r.Host
+
+	return host
+}
+
+func parseForwarded(forwarded string) (addr, proto, host string) {
+	if forwarded == "" {
+		return
+	}
+
+	for _, forwardedPair := range strings.Split(forwarded, ";") {
+		if tv := strings.SplitN(forwardedPair, "=", 2); len(tv) == 2 {
+			token, value := tv[0], tv[1]
+			token = strings.TrimSpace(token)
+			value = strings.TrimSpace(strings.Trim(value, `"`))
+
+			switch strings.ToLower(token) {
+			case "for":
+				addr = value
+			case "proto":
+				proto = value
+			case "host":
+				host = value
+			}
+		}
+	}
+
+	return
 }
